+++ examples/embed/CHANGELOG.md
@@ -1,5 +1,12 @@
 # example-embed
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @tts2go/embed@1.0.0
+
 ## 0.1.1
 
 ### Patch Changes