+++ examples/vanilla/CHANGELOG.md
@@ -1,5 +1,12 @@
 # example-vanilla
 
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @tts2go/vanilla@1.0.0
+
 ## 0.1.0
 
 ### Minor Changes