+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tts2go/core
 
+## 1.0.0
+
+### Major Changes
+
+- adds streaming capability for projects that are configured with it
+
 ## 0.10.0
 
 ### Minor Changes